requirements.txt
requests
beautifulsoup4
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-chroma
chromadb
python-dotenv